feat: add request id and expose context hydrate from metadata
