+++ grpc-context-utils/src/test/java/org/hypertrace/core/grpcutils/context/RequestContextTest.java
@@ -3,9 +3,11 @@ package org.hypertrace.core.grpcutils.context;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import com.google.common.collect.ImmutableList;
+import io.grpc.Metadata;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 /** Unit tests for {@link RequestContext} and utility methods in it. */
@@ -66,5 +68,82 @@ public class RequestContextTest {
     requestContext.add("authorization", "Bearer " + jwt);
     List<String> actualRoles = requestContext.getRoles("roles");
     assertEquals(expectedRoles, actualRoles);
+  }
+
+  @Test
+  public void testMetadataKeys() {
+    Metadata metadata = new Metadata();
+    metadata.put(
+        Metadata.Key.of("authorization", Metadata.ASCII_STRING_MARSHALLER),
+        "Bearer Some-bearer-auth");
+    metadata.put(RequestContextConstants.TENANT_ID_METADATA_KEY, "test-tenant-id");
+
+    RequestContext requestContext = RequestContext.fromMetadata(metadata);
+
+    Assertions.assertEquals(2, requestContext.getAll().size());
+    // GRPC Metadata keys are lower cased during creation.
+    Assertions.assertEquals("Bearer Some-bearer-auth", requestContext.get("authorization").get());
+    Assertions.assertEquals(
+        "test-tenant-id",
+        requestContext.get(RequestContextConstants.TENANT_ID_METADATA_KEY.name()).get());
+
+    metadata = new Metadata();
+    metadata.put(
+        Metadata.Key.of("authorization", Metadata.ASCII_STRING_MARSHALLER),
+        "Bearer Some-bearer-auth");
+    metadata.put(RequestContextConstants.TENANT_ID_METADATA_KEY, "test-tenant-id");
+    metadata.put(
+        Metadata.Key.of("x-some-other-header", Metadata.ASCII_STRING_MARSHALLER),
+        "Some-other-header-val");
+
+    requestContext = RequestContext.fromMetadata(metadata);
+
+    Assertions.assertEquals(2, requestContext.getAll().size());
+    Assertions.assertEquals("Bearer Some-bearer-auth", requestContext.get("authorization").get());
+    Assertions.assertEquals(
+        "test-tenant-id",
+        requestContext.get(RequestContextConstants.TENANT_ID_METADATA_KEY.name()).get());
+
+    // Test that header keys are lowercased
+    metadata = new Metadata();
+    metadata.put(
+        Metadata.Key.of("Authorization", Metadata.ASCII_STRING_MARSHALLER),
+        "Bearer Some-bearer-auth-2");
+    metadata.put(
+        Metadata.Key.of("X-tenant-Id", Metadata.ASCII_STRING_MARSHALLER), "test-tenant-id-2");
+    metadata.put(
+        Metadata.Key.of("X-Some-Other-Header", Metadata.ASCII_STRING_MARSHALLER),
+        "Some-other-header-val-2");
+
+    requestContext = RequestContext.fromMetadata(metadata);
+
+    Assertions.assertEquals(2, requestContext.getAll().size());
+    Assertions.assertEquals("Bearer Some-bearer-auth-2", requestContext.get("authorization").get());
+    Assertions.assertEquals(
+        "test-tenant-id-2",
+        requestContext.get(RequestContextConstants.TENANT_ID_METADATA_KEY.name()).get());
+
+    metadata = new Metadata();
+    metadata.put(
+        Metadata.Key.of("Authorization", Metadata.ASCII_STRING_MARSHALLER),
+        "Bearer Some-bearer-auth-3");
+    metadata.put(
+        Metadata.Key.of("X-tenant-Id", Metadata.ASCII_STRING_MARSHALLER), "test-tenant-id-3");
+    metadata.put(
+        Metadata.Key.of("X-Some-Other-Header", Metadata.ASCII_STRING_MARSHALLER),
+        "Some-other-header-val-3");
+    metadata.put(
+        Metadata.Key.of("grpc-trace-bin", Metadata.BINARY_BYTE_MARSHALLER),
+        "AAARf5ZpQwlN/8FVe1axOPlaAQIdRU/Y8j0LAgE".getBytes());
+
+    requestContext = RequestContext.fromMetadata(metadata);
+
+    Assertions.assertEquals(3, requestContext.getAll().size());
+    Assertions.assertEquals("Bearer Some-bearer-auth-3", requestContext.get("authorization").get());
+    Assertions.assertEquals(
+        "test-tenant-id-3",
+        requestContext.get(RequestContextConstants.TENANT_ID_METADATA_KEY.name()).get());
+    Assertions.assertEquals(
+        "AAARf5ZpQwlN/8FVe1axOPlaAQIdRU/Y8j0LAgE", requestContext.get("grpc-trace-bin").get());
   }
 }
